+++ thermion_dart/lib/src/bindings/src/thermion_dart_js_interop.g.dart
@@ -301,7 +301,7 @@ extension type GeneratedBindings(NativeLibrary _) implements JSObject {
     Pointer<TGltfAssetLoader> tAssetLoader,
     Pointer<TNameComponentManager> tNameComponentManager,
     Pointer<TFilamentAsset> tFilamentAsset,
-    bool rebuildVertices,
+    int vertexBufferMode,
   );
   external Pointer<TFilamentAsset> _SceneAsset_getFilamentAsset(Pointer<TSceneAsset> tSceneAsset);
   external int _SceneAsset_getType(Pointer<TSceneAsset> tSceneAsset);
@@ -1362,7 +1362,7 @@ extension type GeneratedBindings(NativeLibrary _) implements JSObject {
     Pointer<TGltfAssetLoader> tAssetLoader,
     Pointer<TNameComponentManager> tNameComponentManager,
     Pointer<TFilamentAsset> tFilamentAsset,
-    bool rebuildVertices,
+    int vertexBufferMode,
     Pointer<NativeFunction<void Function(PointerClass<TSceneAsset>)>> onComplete,
   );
   external void _SceneAsset_createFromBuffersRenderThread(
@@ -3342,14 +3342,14 @@ Pointer<TSceneAsset> SceneAsset_createFromFilamentAsset(
   Pointer<TGltfAssetLoader> tAssetLoader,
   Pointer<TNameComponentManager> tNameComponentManager,
   Pointer<TFilamentAsset> tFilamentAsset,
-  bool rebuildVertices,
+  int vertexBufferMode,
 ) {
   final result = GeneratedBindings.instance._SceneAsset_createFromFilamentAsset(
     tEngine.cast(),
     tAssetLoader.cast(),
     tNameComponentManager.cast(),
     tFilamentAsset.cast(),
-    rebuildVertices,
+    vertexBufferMode,
   );
   return Pointer<TSceneAsset>(result);
 }
@@ -6176,7 +6176,7 @@ void SceneAsset_createFromFilamentAssetRenderThread(
   Pointer<TGltfAssetLoader> tAssetLoader,
   Pointer<TNameComponentManager> tNameComponentManager,
   Pointer<TFilamentAsset> tFilamentAsset,
-  bool rebuildVertices,
+  int vertexBufferMode,
   Pointer<NativeFunction<void Function(Pointer<TSceneAsset>)>> onComplete,
 ) {
   final result = GeneratedBindings.instance._SceneAsset_createFromFilamentAssetRenderThread(
@@ -6184,7 +6184,7 @@ void SceneAsset_createFromFilamentAssetRenderThread(
     tAssetLoader.cast(),
     tNameComponentManager.cast(),
     tFilamentAsset.cast(),
-    rebuildVertices,
+    vertexBufferMode,
     onComplete.cast(),
   );
   return result;
@@ -10241,6 +10241,12 @@ final class TFilamentAsset extends Struct {
   static Pointer<TFilamentAsset> stackAlloc() {
     return Pointer<TFilamentAsset>(NativeLibrary.instance.stackAlloc<TFilamentAsset>(0));
   }
+}
+
+sealed class TVertexBufferMode {
+  static const VERTEX_BUFFER_MODE_ORIGINAL = 0;
+  static const VERTEX_BUFFER_MODE_UNWELDED = 1;
+  static const VERTEX_BUFFER_MODE_EDITABLE = 2;
 }
 
 sealed class TSceneAssetType {