+++ thermion_dart/native/include/c_api/ThermionDartRenderThreadApi.h
@@ -161,7 +161,7 @@ namespace thermion
             TGltfAssetLoader *tAssetLoader,
             TNameComponentManager *tNameComponentManager,
             TFilamentAsset *tFilamentAsset,
-            bool rebuildVertices,
+            enum TVertexBufferMode vertexBufferMode,
             void (*onComplete)(TSceneAsset *)
         );
         EMSCRIPTEN_KEEPALIVE void SceneAsset_createFromBuffersRenderThread(